xcb: Don't finish snapshots when they are detached
Some code might own a reference to the snapshot when it is
detached. For this reason, we shouldn't finish the snapshot except
when its reference count drops to zero.

This avoids destroying source patterns which get evicted from the
cache while acquiring the mask.

Fixes xcb-stress-cache.

Big "thank you" to Andrea Canciani for helping in figuring this one out.

Signed-off-by: Uli Schlachter <psychon@znc.in>