test: Fix surface-pattern-scale-down-extend
This test creates a similar surface using cairo_push_group () and
assumes its size will be w x h. This is not true when testing similar
surfaces with "cairo-test-suite -a", because they are bigger than the
actual ROI, which corresponds to the rectangle (0,0) - (w,h).

Clipping to the ROI guarantees that the surface created by
cairo_push_group () always has the same size.

Fixes surface-pattern-scale-down-extend.