Revert "xcb: Never return NULL from create_similar"
This reverts commit f6e6891b59656c944a7e758a53e07ed5f29f5c39.

Returning NULL in these cases is actually the right thing to do, because the
caller expects and handles this case.

cairo_surface_create_similar() won't return NULL due to this, because
_cairo_surface_create_similar_solid() will fall back to an image surface.

I think I missed that part when I wrote the bad commit.

This was found via the test suite's "zero-mask" test.

Signed-off-by: Uli Schlachter <psychon@znc.in>