scaled-font: Key the cache on the orignal font_face
We check the incoming scaled font using the original font fce, so we
need to also store it in the cache using the same face, and not the
resolved font_face (which will remain the same unless the fontconfig
configuration is updated).

Hides the quadratic behaviour of font retrieval in recent cairo-traces.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>