type1-subset: Remove FT dependency part 1
Remove cairo_type1_font_subset_get_glyph_names_and_widths().

The glyph names are read from the font file instead of via FT. The
charstrings are parsed to extract the glyph widths.

A new font backend function, index_to_glyph_name, has been added for
obtaining the glyph name for a given glyph index. This function is
supplied with the array of glyph names and a glyph index and is
required to return the array index of the glyph name corresponding to
the glyph index.

The reason for passing in the array of glyph names is that:

1) On windows there is no API for accessing glyph names so we will
use knowledge of how the glyphs in a Type 1 font are numbered to
perform name lookup.

2) We can also use knowledge of how FT assigns the glyph numbers in a
Type 1 font to optimize the name lookup.