record: Recording surfaces need an explicit clear
As a recording surface will be replayed onto a destination target, it is
unknown at the time of creation as to whether or not the target is
clear. So we need to make sure to record any initial clear in order
to replay that clear onto the destination. (If the destination is clear
anyway it will be suppressed.)

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>